Create the ViewInBlockExplorer component
