+++ solidity/dashboard/src/components/AuthorizeContracts.jsx
@@ -2,7 +2,7 @@ import React, { useCallback } from "react"
 import AddressShortcut from "./AddressShortcut"
 import { SubmitButton } from "./Button"
 import { DataTable, Column } from "./DataTable"
-import ViewAddressInBlockExplorer from "./ViewAddressInBlockExplorer"
+import { ViewAddressInBlockExplorer } from "./ViewInBlockExplorer"
 import { displayAmount } from "../utils/token.utils"
 import StatusBadge, { BADGE_STATUS } from "./StatusBadge"
 import SpeechBubbleTooltip from "./SpeechBubbleTooltip"
